[Chore #118744211] Refactor UserDashboardController
- Delegate video specific operations to `VideoController`
- Update video CRUD routes in `route.php` to point to `VideoController`
- Migrate some tests in `UserDashoardTest.php` into `VideoControllerTest.php`
